from: mdk369@users.sf.net make_lpd_conf used for generating html files prints timestamps and causes build to be unreproducible**
Hello, I'm submitting this change in order to further the reproducible builds agenda - which aims to
ensure all projects built remain bit-by-bit identical uniformly throughout the package.

In this case, a date being printed within make_lpd_conf.in that would be generated in an html output
was causing this build to not be building in a reproducible fashion.

* Since this is not a strict requierement we like to help and remove the date